refactor: split script for easier maintanability and readability
